auth tutorial 15 - authorized requests

+++ frontend/src/components/WorkoutDetails.js
@@ -1,3 +1,4 @@
+import { useAuthContext } from "../hooks/useAuthContext";
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 
 // date fns (not installed with npm does not function correctly)
@@ -5,10 +6,18 @@ import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 
 const WorkoutDetails = ({ workout }) => {
     const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
 
     const handleDelete = async () => {
+        if(!user){
+            return
+        }
+
         const response = await fetch("http://localhost:4000/api/workouts/" + workout._id, {
-            method: "DELETE"
+            method: "DELETE",
+            headers:{
+                "Authorization": user.token
+            }
         })
 
         const json = await response.json()
